Baicheng

+++ app/src/main/java/com/teambald/cse442_project_team_bald/TabsController/SwipeActionHandler.java
@@ -1,7 +1,18 @@
 package com.teambald.cse442_project_team_bald.TabsController;
+import android.util.Log;
+import android.view.View;
+
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.teambald.cse442_project_team_bald.Objects.RecordingItem;
+import com.teambald.cse442_project_team_bald.R;
+import com.teambald.cse442_project_team_bald.Fragments.CloudFragment;
+import com.teambald.cse442_project_team_bald.Fragments.RecordingListFragment;
+
+import java.io.File;
+import java.util.ArrayList;
 
 public class SwipeActionHandler extends ItemTouchHelper.SimpleCallback {
 
@@ -16,10 +27,27 @@ public class SwipeActionHandler extends ItemTouchHelper.SimpleCallback {
      */
     private RecordingListAdapter mAdapter;
 
+    //fragmentId = 0 -> recording frag
+    //fragmentId = 1 -> cloud frag
+    private int fragmentId = -1;
 
-    public SwipeActionHandler(RecordingListAdapter adapter) {
+    private static final String TAG = "SwipeActionHandler";
+
+    private CloudFragment cloudFragment;
+    private RecordingListFragment recordingListFragment;
+
+    public SwipeActionHandler(RecordingListAdapter adapter,int fragment,CloudFragment frag) {
         super(0,ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT);
         mAdapter = adapter;
+        fragmentId = fragment;
+        cloudFragment = frag;
+    }
+
+    public SwipeActionHandler(RecordingListAdapter adapter,int fragment,RecordingListFragment frag) {
+        super(0,ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT);
+        mAdapter = adapter;
+        fragmentId = fragment;
+        recordingListFragment = frag;
     }
 
     @Override
@@ -30,7 +58,51 @@ public class SwipeActionHandler extends ItemTouchHelper.SimpleCallback {
     @Override
     public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
         int position = viewHolder.getAdapterPosition();
-        mAdapter.deleteItem(position);
-        mAdapter.notifyDataSetChanged();
+        ArrayList<RecordingItem> recordingList = mAdapter.getmDataset();
+        switch (fragmentId)
+        {
+            case 0:
+                if(direction == ItemTouchHelper.LEFT)
+                {
+                    //Delete file
+                    Log.d(TAG,"Deleting file");
+                    mAdapter.deleteItem(position);
+                    mAdapter.notifyDataSetChanged();
+                }
+                else if (direction == ItemTouchHelper.RIGHT)
+                {
+                    //Upload file
+                    mAdapter.notifyDataSetChanged();
+                    RecordingItem item = recordingList.get(position);
+                    File file = item.getAudio_file();
+                    Log.d(TAG,"Uploading file: "+file.getAbsolutePath());
+                }
+                //
+                break;
+            case 1:
+                if(direction == ItemTouchHelper.LEFT)
+                {
+                    //Download file
+                    mAdapter.notifyDataSetChanged();
+                    RecordingItem item = recordingList.get(position);
+                    File file = item.getAudio_file();
+                    Log.d(TAG,"Downloading file: "+file.getAbsolutePath());
+                }
+                else if (direction == ItemTouchHelper.RIGHT)
+                {
+                    //Delete file
+                    Log.d(TAG,"Deleting file");
+                    mAdapter.deleteItem(position);
+                    mAdapter.notifyDataSetChanged();
+                }
+                //
+                break;
+            case -1:
+                Log.d(TAG,"swipe error, fail to initialize fragment id number");
+                break;
+            default:
+                Log.d(TAG,"swipe error, unknown case number");
+
+        }
     }
 }
